src/requirements.txt
azure-core>=1.26.0
azure-ai-documentintelligence>=1.0.0
azure-storage-blob>=12.14.0
streamlit>=1.28.0
python-dotenv>=1.0.0